+++ Code/Controllers/LYRUIConversationListViewController.m
@@ -124,6 +124,10 @@ NSString *const LYRUIConversationTableViewAccessibilityIdentifier = @"Conversati
 - (void)viewWillAppear:(BOOL)animated
 {
     [super viewWillAppear:animated];
+    // Hide the search bar
+    CGFloat contentOffset = self.tableView.contentOffset.y + self.searchBar.frame.size.height;
+    [self.tableView setContentOffset:CGPointMake(0, contentOffset)];
+    
     [self setupConversationDataSource];
     if (!self.hasAppeared) {
         [self.tableView registerClass:self.cellClass forCellReuseIdentifier:LYRUIConversationCellReuseIdentifier];